refactor: move pagination to product list component
